requirements.txt
blinker==1.8.2
click==8.1.7
ffmpeg-python==0.2.0
Flask==3.0.3
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.4
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.3
